Silence compiler warning: int vs unsigned int compare

+++ src/gui/sdlmain.cpp
@@ -8623,7 +8623,7 @@ bool video_frameskip_common_menu_callback(DOSBoxMenu * const menu,DOSBoxMenu::it
     SetVal("render", "frameskip", tmp1);
     for (unsigned int i=0;i<=10;i++) {
         sprintf(tmp2,"frameskip_%u",i);
-        mainMenu.get_item(tmp2).check(f==i).refresh_item(mainMenu);
+        mainMenu.get_item(tmp2).check((unsigned int)f==i).refresh_item(mainMenu);
     }
     return true;
 }
